azure-ai-vision/requirements.txt
azure-ai-vision
pdf2image
pillow
python-dotenv
